types: move storage-related types to types/storage.py
