+++ subiquity/common/apidef.py
@@ -26,7 +26,6 @@ from subiquity.common.types import (
     AdAdminNameValidation,
     AdConnectionInfo,
     AdDomainNameValidation,
-    AddPartitionV2,
     AdJoinResult,
     AdPasswordValidation,
     AnyStep,
@@ -35,12 +34,9 @@ from subiquity.common.types import (
     CasperMd5Results,
     Change,
     CodecsData,
-    Disk,
     DriversPayload,
     DriversResponse,
     ErrorReportRef,
-    GuidedChoiceV2,
-    GuidedStorageResponseV2,
     IdentityData,
     KeyboardSetting,
     KeyboardSetup,
@@ -50,11 +46,9 @@ from subiquity.common.types import (
     MirrorPost,
     MirrorPostResponse,
     MirrorSelectionFallback,
-    ModifyPartitionV2,
     NetworkStatus,
     OEMResponse,
     PackageInstallState,
-    ReformatDisk,
     RefreshStatus,
     ShutdownMode,
     SnapInfo,
@@ -63,8 +57,6 @@ from subiquity.common.types import (
     SourceSelectionAndSetting,
     SSHData,
     SSHFetchIdResponse,
-    StorageResponse,
-    StorageResponseV2,
     TimeZoneInfo,
     UbuntuProCheckTokenAnswer,
     UbuntuProGeneralInfo,
@@ -74,6 +66,16 @@ from subiquity.common.types import (
     UPCSWaitResponse,
     UsernameValidation,
     ZdevInfo,
+)
+from subiquity.common.types.storage import (
+    AddPartitionV2,
+    Disk,
+    GuidedChoiceV2,
+    GuidedStorageResponseV2,
+    ModifyPartitionV2,
+    ReformatDisk,
+    StorageResponse,
+    StorageResponseV2,
 )
 from subiquitycore.models.network import (
     BondConfig,
